RF: Move eyetracking to its own module under psychopy
