+++ psychopy/eyetracking/__init__.py
@@ -1,0 +1,3 @@
+from .record import EyetrackerControl
+from .calibrate import EyetrackerCalibration
+from .roi import ROI
