+++ psychopy/eyetracking/record.py
@@ -1,0 +1,39 @@
+from psychopy.constants import STARTED, NOT_STARTED, PAUSED, STOPPED, FINISHED
+
+
+class EyetrackerControl:
+    def __init__(self, server, tracker=None):
+        if tracker is None:
+            tracker = server.getDevice('tracker')
+        self.server = server
+        self.tracker = tracker
+        self._status = NOT_STARTED
+
+    @property
+    def status(self):
+        return self._status
+
+    @status.setter
+    def status(self, value):
+        old = self._status
+        new = self._status = value
+        # Skip if there's no change
+        if new == old:
+            return
+        # Start recording if set to STARTED
+        if new in (STARTED,):
+            if old in (NOT_STARTED, STOPPED, FINISHED):
+                # If was previously at a full stop, clear events before starting again
+                self.server.clearEvents()
+            # Start recording
+            self.tracker.setRecordingState(True)
+        # Stop recording if set to any stop constants
+        if new in (NOT_STARTED, PAUSED, STOPPED, FINISHED):
+            self.tracker.setRecordingState(False)
+
+    @property
+    def pos(self):
+        return self.tracker.getPosition()
+
+    def getPos(self):
+        return self.pos
