+++ psychopy/hardware/__init__.py
@@ -10,14 +10,13 @@ import sys
 import glob
 from itertools import chain
 from psychopy import logging
-from . import eyetracker
 
 try:
     from collections.abc import Iterable
 except ImportError:
     from collections import Iterable
 
-__all__ = ['forp', 'cedrus', 'minolta', 'pr', 'crs', 'iolab', 'eyetracker']
+__all__ = ['forp', 'cedrus', 'minolta', 'pr', 'crs', 'iolab']
 
 
 def getSerialPorts():
