+++ psychopy/visual/__init__.py
@@ -39,7 +39,6 @@ from .form import Form
 from .brush import Brush
 from .textbox2.textbox2 import TextBox2
 from .button import ButtonStim
-from .roi import ROI
 from .target import TargetStim
 # window, should always be loaded first
 from .window import Window, getMsPerFrame, openWindows
